design is complete to 75%

+++ lib/features/weather_info/data/models/current_model.dart
@@ -50,7 +50,7 @@ class CurrentModel extends CurrentEntity {
       pressure: json['pressure'],
       humidity: json['humidity'],
       dewPoint: json['dew_point'].toDouble(),
-      uvi: json['uvi'],
+      uvi: json['uvi'].toDouble(),
       clouds: json['clouds'],
       visibility: json['visibility'],
       windSpeed: json['wind_speed'].toDouble(),
