+++ lib/features/weather_info/domain/entities/minutely_entity.dart
@@ -1,6 +1,6 @@
 class MinutelyEntity {
   final int dt;
-  final int precipitation;
+  final double precipitation;
 
   MinutelyEntity({
     required this.dt,
